Relevant BUGIDs:
Purpose of commit:  bugfix, release

Commit summary:
---------------

2007-07-06  Thorsten Kukuk  <kukuk@thkukuk.de>

        * release version 0.99.8.0

        * configure.in: Check for audit_log_acct_message instead of
        audit_log_user_message.
        * libpam/pam_audit.c: Use audit_log_acct_message.
        Based on patch from Mark J Cox <mjc@redhat.com>.
        * libpam/Makefile.am: Bump version number of libpam.

        * modules/pam_umask/pam_umask.c (set_umask): mode_t is 32bit,
        not 64bit.

        * xtests/tst-pam_limits1.c: Fix printf arguments.

        * po/*.po: Merge po files with latest code changes.

+++ modules/pam_umask/pam_umask.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2005, 2006 Thorsten Kukuk <kukuk@thkukuk.de>
+ * Copyright (c) 2005, 2006, 2007 Thorsten Kukuk <kukuk@thkukuk.de>
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -181,7 +181,7 @@ set_umask (const char *value)
 
   mask = strtoul (value, &endptr, 8) & 0777;
   if (((mask == 0) && (value_orig == endptr)) ||
-      ((mask == ULONG_MAX) && (errno == ERANGE)))
+      ((mask == UINT_MAX) && (errno == ERANGE)))
     return;
   umask (mask);
   return;
